CRYPTO: fixed shared key and nonce
We now use a fixed shared key and hardcoded nonce, to further reduce
data transmitted over the mesh.

Shared key is derived for each message dynamically, but reverse-lookup
of "sender_id" (e.g. a short GID) in the network Router. This lookup
byt the receiver will return th (sender) pubkey, which can be used to
remotely generate the shared key.

We prepend this 1 byte "sender_id" to each encrypted message.

Signed-off-by: willcl-ark <will8clark@gmail.com>